SAAS_PLATFORM/backend/requirements.txt
fastapi
uvicorn
pydantic
openai
huggingface_hub
python-dotenv
supabase
